Refactor FXIOS-6595 [v116] Reorganize privacy settings (#14869)
* Reorganize privacy settings

* Indentation

* Fix one identifier
