+++ Client/Frontend/Settings/Main/AppSettingsOptions.swift
@@ -223,237 +223,3 @@ class OpenSupportPageSetting: Setting {
         }
     }
 }
-
-class LoginsSetting: Setting {
-    let profile: Profile
-    var tabManager: TabManager!
-    private let appAuthenticator: AppAuthenticationProtocol
-    weak var navigationController: UINavigationController?
-    weak var settings: AppSettingsTableViewController?
-
-    override var accessoryView: UIImageView? { return SettingDisclosureUtility.buildDisclosureIndicator(theme: theme) }
-
-    override var accessibilityIdentifier: String? { return "Logins" }
-
-    init(settings: SettingsTableViewController,
-         delegate: SettingsDelegate?,
-         appAuthenticator: AppAuthenticationProtocol = AppAuthenticator()) {
-        self.profile = settings.profile
-        self.tabManager = settings.tabManager
-        self.appAuthenticator = appAuthenticator
-        self.navigationController = settings.navigationController
-        self.settings = settings as? AppSettingsTableViewController
-
-        super.init(
-            title: NSAttributedString(
-                string: .Settings.Passwords.Title,
-                attributes: [NSAttributedString.Key.foregroundColor: settings.themeManager.currentTheme.colors.textPrimary]
-            ),
-            delegate: delegate
-        )
-    }
-
-    func deselectRow () {
-        if let selectedRow = self.settings?.tableView.indexPathForSelectedRow {
-            self.settings?.tableView.deselectRow(at: selectedRow, animated: true)
-        }
-    }
-
-    override func onClick(_: UINavigationController?) {
-        deselectRow()
-
-        guard let navController = navigationController else { return }
-        let navigationHandler: (_ url: URL?) -> Void = { url in
-            guard let url = url else { return }
-            UIWindow.keyWindow?.rootViewController?.dismiss(animated: true, completion: nil)
-            self.delegate?.settingsOpenURLInNewTab(url)
-        }
-
-        if appAuthenticator.canAuthenticateDeviceOwner() {
-            if LoginOnboarding.shouldShow() {
-                let loginOnboardingViewController = LoginOnboardingViewController(profile: profile, tabManager: tabManager)
-
-                loginOnboardingViewController.doneHandler = {
-                    loginOnboardingViewController.dismiss(animated: true)
-                }
-
-                loginOnboardingViewController.proceedHandler = {
-                    LoginListViewController.create(
-                        didShowFromAppMenu: false,
-                        authenticateInNavigationController: navController,
-                        profile: self.profile,
-                        webpageNavigationHandler: navigationHandler
-                    ) { loginsVC in
-                        guard let loginsVC = loginsVC else { return }
-                        navController.pushViewController(loginsVC, animated: true)
-                        // Remove the onboarding from the navigation stack so that we go straight back to settings
-                        navController.viewControllers.removeAll { viewController in
-                            viewController == loginOnboardingViewController
-                        }
-                    }
-                }
-
-                navigationController?.pushViewController(loginOnboardingViewController, animated: true)
-
-                LoginOnboarding.setShown()
-            } else {
-                LoginListViewController.create(
-                    didShowFromAppMenu: false,
-                    authenticateInNavigationController: navController,
-                    profile: profile,
-                    webpageNavigationHandler: navigationHandler
-                ) { loginsVC in
-                    guard let loginsVC = loginsVC else { return }
-                    navController.pushViewController(loginsVC, animated: true)
-                }
-            }
-        } else {
-            let viewController = DevicePasscodeRequiredViewController()
-            viewController.profile = profile
-            viewController.tabManager = tabManager
-            navigationController?.pushViewController(viewController, animated: true)
-        }
-    }
-}
-
-class ContentBlockerSetting: Setting {
-    let profile: Profile
-    var tabManager: TabManager!
-    override var accessoryView: UIImageView? { return SettingDisclosureUtility.buildDisclosureIndicator(theme: theme) }
-    override var accessibilityIdentifier: String? { return "TrackingProtection" }
-
-    override var status: NSAttributedString? {
-        let isOn = profile.prefs.boolForKey(ContentBlockingConfig.Prefs.EnabledKey) ?? ContentBlockingConfig.Defaults.NormalBrowsing
-
-        if isOn {
-            let currentBlockingStrength = profile
-                .prefs
-                .stringForKey(ContentBlockingConfig.Prefs.StrengthKey)
-                .flatMap(BlockingStrength.init(rawValue:)) ?? .basic
-            return NSAttributedString(string: currentBlockingStrength.settingStatus)
-        } else {
-            return NSAttributedString(string: .Settings.Homepage.Shortcuts.ToggleOff)
-        }
-    }
-
-    override var style: UITableViewCell.CellStyle { return .value1 }
-
-    init(settings: SettingsTableViewController) {
-        self.profile = settings.profile
-        self.tabManager = settings.tabManager
-        super.init(title: NSAttributedString(string: .SettingsTrackingProtectionSectionName, attributes: [NSAttributedString.Key.foregroundColor: settings.themeManager.currentTheme.colors.textPrimary]))
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        let viewController = ContentBlockerSettingViewController(prefs: profile.prefs)
-        viewController.profile = profile
-        viewController.tabManager = tabManager
-        navigationController?.pushViewController(viewController, animated: true)
-    }
-}
-
-class ClearPrivateDataSetting: Setting {
-    let profile: Profile
-    var tabManager: TabManager!
-
-    override var accessoryView: UIImageView? { return SettingDisclosureUtility.buildDisclosureIndicator(theme: theme) }
-
-    override var accessibilityIdentifier: String? { return "ClearPrivateData" }
-
-    init(settings: SettingsTableViewController) {
-        self.profile = settings.profile
-        self.tabManager = settings.tabManager
-
-        let clearTitle: String = .SettingsDataManagementSectionName
-        super.init(title: NSAttributedString(string: clearTitle, attributes: [NSAttributedString.Key.foregroundColor: settings.themeManager.currentTheme.colors.textPrimary]))
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        let viewController = ClearPrivateDataTableViewController()
-        viewController.profile = profile
-        viewController.tabManager = tabManager
-        navigationController?.pushViewController(viewController, animated: true)
-    }
-}
-
-class AutofillCreditCardSettings: Setting, FeatureFlaggable {
-    private let profile: Profile
-    private let appAuthenticator: AppAuthenticationProtocol
-    weak var navigationController: UINavigationController?
-    weak var settings: AppSettingsTableViewController?
-    override var accessoryView: UIImageView? { return SettingDisclosureUtility.buildDisclosureIndicator(theme: theme) }
-    override var accessibilityIdentifier: String? { return "AutofillCreditCard" }
-
-    init(settings: SettingsTableViewController,
-         appAuthenticator: AppAuthenticationProtocol = AppAuthenticator()) {
-        self.profile = settings.profile
-        self.appAuthenticator = appAuthenticator
-        self.navigationController = settings.navigationController
-        self.settings = settings as? AppSettingsTableViewController
-
-        super.init(
-            title: NSAttributedString(
-                string: .SettingsAutofillCreditCard,
-                attributes: [NSAttributedString.Key.foregroundColor: settings.themeManager.currentTheme.colors.textPrimary]
-            )
-        )
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        TelemetryWrapper.recordEvent(category: .action, method: .tap, object: .creditCardAutofillSettings)
-        let viewModel = CreditCardSettingsViewModel(profile: profile)
-        let viewController = CreditCardSettingsViewController(
-            creditCardViewModel: viewModel)
-
-        guard let navController = navigationController else { return }
-        if appAuthenticator.canAuthenticateDeviceOwner() {
-            AppAuthenticator().authenticateWithDeviceOwnerAuthentication { result in
-                switch result {
-                case .success:
-                    navController.pushViewController(viewController,
-                                                     animated: true)
-                case .failure:
-                    viewController.dismissVC()
-                }
-            }
-        } else {
-            let passcodeViewController = DevicePasscodeRequiredViewController()
-            passcodeViewController.profile = profile
-            navController.pushViewController(passcodeViewController,
-                                             animated: true)
-        }
-    }
-}
-
-class PrivacyPolicySetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .AppSettingsPrivacyPolicy, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override var url: URL? {
-        return URL(string: "https://www.mozilla.org/privacy/firefox/")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-class NotificationsSetting: Setting {
-    override var accessoryView: UIImageView? { return SettingDisclosureUtility.buildDisclosureIndicator(theme: theme) }
-
-    override var accessibilityIdentifier: String? { return AccessibilityIdentifiers.Setting.notifications }
-
-    let profile: Profile
-
-    init(theme: Theme, profile: Profile) {
-        self.profile = profile
-        super.init(title: NSAttributedString(string: .Settings.Notifications.Title,
-                                             attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary]))
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        let viewController = NotificationsSettingsViewController(prefs: profile.prefs, hasAccount: profile.hasAccount())
-        navigationController?.pushViewController(viewController, animated: true)
-    }
-}
